Token JWT fonctionnel

+++ Front/Services/CustomAuthenticationStateProvider.cs
@@ -3,6 +3,8 @@ using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.AspNetCore.Components.Authorization;
 using Microsoft.AspNetCore.Components.Server.ProtectedBrowserStorage;
 using Microsoft.AspNetCore.Mvc;
+using System.Net.Http.Headers;
+using System.Net.Http;
 using System.Security.Claims;
 using System.Security.Principal;
 
@@ -35,6 +37,7 @@ namespace Front.Services
         public async Task<ClaimsPrincipal> Logout()
         {
             await _sessionStorage.DeleteAsync("User");
+            await _sessionStorage.DeleteAsync("jwt");
             _currentUser = new ClaimsPrincipal(new ClaimsIdentity());
 
             NotifyAuthenticationStateChanged(GetAuthenticationStateAsync());
@@ -51,21 +54,32 @@ namespace Front.Services
 
                 using (HttpClient client = new HttpClient())
                 {
-                    // Concaténez l'ID à l'URL de l'API
-                    string urlWithId = $"http://localhost:5000/api/User/{user.Id}";
-
-                    // Envoie la requête HTTP DELETE
-                    HttpResponseMessage response = await client.DeleteAsync(urlWithId);
-
-                    // Vérifie la réponse
-                    if (response.IsSuccessStatusCode)
+                    var token = await _sessionStorage.GetAsync<string>("jwt");
+                    if(token.Success)
                     {
-                        Console.WriteLine($"Suppression réussie pour l'ID {user.Id}");
+                        client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token.Value);
+
+                        // Concaténez l'ID à l'URL de l'API
+                        string urlWithId = $"http://localhost:5000/api/User/{user.Id}";
+
+                        // Envoie la requête HTTP DELETE
+                        HttpResponseMessage response = await client.DeleteAsync(urlWithId);
+
+                        // Vérifie la réponse
+                        if (response.IsSuccessStatusCode)
+                        {
+                            Console.WriteLine($"Suppression réussie pour l'ID {user.Id}");
+                        }
+                        else
+                        {
+                            Console.WriteLine($"Erreur lors de la suppression. Code de statut : {response.StatusCode}");
+                        }
                     }
                     else
                     {
-                        Console.WriteLine($"Erreur lors de la suppression. Code de statut : {response.StatusCode}");
+                        Console.WriteLine("Erreur : Token JWT impossible à récupérer");
                     }
+                    
                 }
             }
